+++ apps/docs/app/(introduction)/layout.tsx
@@ -1,0 +1,15 @@
+import type { ReactNode } from "react";
+import { Header } from "~/components/Header";
+
+export default function IntroductionLayout({
+  children,
+}: {
+  children: ReactNode;
+}) {
+  return (
+    <>
+      <Header startAtCenter />
+      <main className="p-5">{children}</main>
+    </>
+  );
+}
