+++ apps/docs/app/(top-header)/colors/page.tsx
@@ -1,0 +1,3 @@
+export default function Colors() {
+  return "Colors";
+}
