+++ apps/docs/app/(top-header)/components/accordion/page.tsx
@@ -1,0 +1,3 @@
+export default function AccordionPage() {
+  return "Accordion";
+}
