+++ apps/docs/app/(top-header)/components/alert/page.tsx
@@ -1,0 +1,3 @@
+export default function AlertPage() {
+  return "Alert";
+}
