+++ apps/docs/app/(top-header)/components/button/page.tsx
@@ -1,0 +1,3 @@
+export default function ButtonPage() {
+  return "Button";
+}
