+++ apps/docs/app/(top-header)/components/card/page.tsx
@@ -1,0 +1,3 @@
+export default function CardPage() {
+  return "Card";
+}
