+++ apps/docs/app/(top-header)/components/heading/page.tsx
@@ -1,0 +1,3 @@
+export default function HeadingPage() {
+  return "Heading";
+}
