+++ apps/docs/app/(top-header)/components/input/page.tsx
@@ -1,0 +1,3 @@
+export default function InputPage() {
+  return "Input";
+}
