+++ apps/docs/app/(top-header)/components/layout.tsx
@@ -1,0 +1,5 @@
+import type { ReactNode } from "react";
+
+export default function ComponentLayout({ children }: { children: ReactNode }) {
+  return children;
+}
