+++ apps/docs/app/(top-header)/components/modal/page.tsx
@@ -1,0 +1,3 @@
+export default function ModalPage() {
+  return "Modal";
+}
