+++ apps/docs/app/(top-header)/components/sidebar/page.tsx
@@ -1,0 +1,3 @@
+export default function SidebarPage() {
+  return "Sidebar";
+}
