+++ apps/docs/app/(top-header)/components/text/page.tsx
@@ -1,0 +1,3 @@
+export default function TextPage() {
+  return "Text";
+}
