+++ apps/docs/app/(top-header)/layout.tsx
@@ -1,0 +1,11 @@
+import type { ReactNode } from "react";
+import { Header } from "~/components/Header";
+
+export default function TopHeaderLayout({ children }: { children: ReactNode }) {
+  return (
+    <>
+      <Header />
+      <main className="p-5 max-w-prose">{children}</main>
+    </>
+  );
+}
