+++ apps/docs/app/(top-header)/theming/page.tsx
@@ -1,0 +1,3 @@
+export default function Theming() {
+  return "Theming";
+}
